Update README.md
An Electronic Monitoring System (EMS) refers to a technology-based system used for tracking and monitoring the activities of individuals, equipment, or processes. EMS is commonly used in security, law enforcement, workplace monitoring, and even in healthcare. It can involve a combination of hardware and software to collect, analyze, and report data.